+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2021-10-19
+### Added
+- frame counter
+### Removed
+- google analytics
+
 ## [0.2.2] - 2021-10-19
 ### Fixed
 - links in changelog
@@ -22,7 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [HTML5 Boilerplate 8.0](https://html5boilerplate.com)
 
-[Unreleased]: https://github.com/ArtemNikolaev/explosion-js-generator-demo/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/ArtemNikolaev/explosion-js-generator-demo/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/ArtemNikolaev/explosion-js-generator-demo/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/ArtemNikolaev/explosion-js-generator-demo/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/ArtemNikolaev/explosion-js-generator-demo/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/ArtemNikolaev/explosion-js-generator-demo/compare/v0.1.0...v0.2.0